Remove outdated structure list
